# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Inventory App API
  version: 1.0.0
  summary: Internal inventory management service
  description: >
    Inventory management API covering authentication, products, stock flows, alerts,
    and configuration. All endpoints (except login) require a bearer token issued to
    internal users.
servers:
  - url: /api/v1
    description: Primary API entrypoint
...
security:
  - bearerAuth: []
paths:
...
  /users:
    get:
      tags: [Users]
      summary: 사용자 목록
      operationId: listUsers
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/SizeParam'
      responses:
        '200':
          description: 사용자 목록 응답
          content:
            application/json:
              schema:
                type: object
                required: [data, page]
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/User'
                  page:
                    $ref: '#/components/schemas/Page'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      tags: [Users]
      summary: 사용자 생성
      operationId: createUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, name, role]
              properties:
                email:
                  type: string
                  format: email
                name:
                  type: string
                role:
                  $ref: '#/components/schemas/Role'
                permissions:
                  type: array
                  items:
                    $ref: '#/components/schemas/Permission'
      responses:
        '201':
          description: 사용자 생성 완료
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  parameters:
    PageParam:
      name: page
      in: query
      description: 페이지 번호
      schema:
        type: integer
        minimum: 1
        default: 1
    SizeParam:
      name: size
      in: query
      description: 페이지당 항목 수
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 20
  responses:
    BadRequest:
      description: 잘못된 요청
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    Unauthorized:
      description: 인증 필요
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
...
  schemas:
    ApiError:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: object
          nullable: true
    Page:
      type: object
      properties:
        page:
          type: integer
        size:
          type: integer
        total:
          type: integer
      required: [page, size, total]
    Role:
      type: string
      enum: [admin, operator, viewer]
    Permission:
      type: object
      properties:
        resource:
          type: string
          enum:
            [
              dashboard,
              products,
              inbounds,
              outbounds,
              returns,
              settings,
            ]
        read:
          type: boolean
        write:
          type: boolean
      required: [resource, read, write]
    User:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
          format: email
        name:
          type: string
        role:
          $ref: '#/components/schemas/Role'
        permissions:
          type: array
          items:
            $ref: '#/components/schemas/Permission'
      required: [id, email, name, role]
...